+++ src/compas/datastructures/mesh/_mesh.py
@@ -34,8 +34,22 @@ class Mesh(BaseMesh):
     Examples
     --------
     >>> mesh = Mesh.from_polyhedron(6)
+
     """
-    bounding_box = mesh_bounding_box
+
+    def bounding_box(self):
+        """Compute the axis-aligned bounding box of the mesh.
+
+        Returns
+        -------
+        box: tuple of 8 points.
+            The corners of the bounding box.
+            The first 4 points are the corners of the bottom face, in counter clockwise direction wrt the positive Z-axis.
+            The last 4 points are the corners of the top face, in counter clockwise direction wrt the positive Z-axis.
+
+        """
+        return mesh_bounding_box(self)
+
     bounding_box_xy = mesh_bounding_box_xy
     collapse_edge = mesh_collapse_edge
     connected_components = mesh_connected_components
@@ -54,20 +68,6 @@ class Mesh(BaseMesh):
     transformed = mesh_transformed
     unify_cycles = mesh_unify_cycles
     quads_to_triangles = mesh_quads_to_triangles
-
-    def transform_numpy(self, M):
-        from compas.datastructures.mesh.transformations_numpy import mesh_transform_numpy
-        mesh_transform_numpy(self, M)
-
-    # def to_trimesh(self):
-    #     # convert to mesh with only triangle faces
-    #     # provides options that define the rules for triangulation
-    #     # for use with trimesh-specific algorithms
-    #     # provide option to use numpy for storage of vertices and faces
-    #     pass
-
-    # def to_quadmesh(self):
-    #     pass
 
 
 # =============================================================================