+++ src/compas/geometry/__init__.py
@@ -5,21 +5,11 @@ geometry
 
 .. currentmodule:: compas.geometry
 
+Classes
+=======
 
 Primitives
-==========
-
-Base Classes
-------------
-
-.. autosummary::
-    :toctree: generated/
-    :nosignatures:
-
-    Primitive
-
-Classes
--------
+----------
 
 .. autosummary::
     :toctree: generated/
@@ -37,58 +27,8 @@ Classes
     Vector
 
 
-The following representations of primitives can be used
-interchangeably as input in methods and functions.
-The representations using native Python objects also correspond to the required input
-parameters of the default constructor functions of the corresponding COMPAS objects.
-
-============  ==============================   ============
-Object        COMPAS                           Python
-------------  ------------------------------   ------------
-point         Point(float, float, float)       [float, float, float]
-vector        Vector(float, float, float)      [float, float, float]
-line          Line(point, point)               [[float, float, float], [float, float, float]]
-plane         Plane(point, vector)             [[float, float, float], [float, float, float]]
-circle        Circle(plane, float)             [[[float, float, float], [float, float, float]], float]
-polygon       Polygon(points)                  [[float, float, float], ... [float, float, float]]
-polyline      Polyline(points)                 [[float, float, float], ... [float, float, float]]
-ellipse       Ellipse(plane, float, float)     [[[float, float, float], [float, float, float]], float, float]
-frame         Frame(point, vector,  vector)    [[float, float, float], [float, float, float], [float, float, float]]
-============  ==============================   ============
-
-COMPAS primitives also support indexing, assignment, and iteration according to the above
-equivalency.
-
-::
-
-    >>> a = Point(0, 0, 0)
-    >>> x, y, z = a
-    >>> x = a[0]
-    >>> a[0] = 0
-
-::
-
-    >>> plane = Plane(Point(0, 0, 0), Vector(0, 0, 1))
-    >>> a, n = plane
-    >>> x, y, z = n
-    >>> plane[0] = Point(1, 0, 0)
-
-
 Shapes
-======
-
-Base Classes
-------------
-
-.. autosummary::
-    :toctree: generated/
-    :nosignatures:
-
-    Shape
-
-
-Classes
--------
+------
 
 .. autosummary::
     :toctree: generated/
@@ -103,8 +43,27 @@ Classes
     Torus
 
 
+Transformations
+---------------
+
+.. autosummary::
+    :toctree: generated/
+    :nosignatures:
+
+    Projection
+    Reflection
+    Rotation
+    Scale
+    Shear
+    Transformation
+    Translation
+
+
+Functions
+=========
+
 Predicates 2D
-=============
+-------------
 
 .. autosummary::
     :toctree: generated/
@@ -124,7 +83,7 @@ Predicates 2D
 
 
 Predicates 3D
-=============
+-------------
 
 .. autosummary::
     :toctree: generated/
@@ -134,7 +93,6 @@ Predicates 3D
     is_coplanar
     is_point_in_halfspace
     is_point_in_polyhedron
-    is_point_in_sphere
     is_point_on_line
     is_point_on_plane
     is_point_on_polyline
@@ -143,23 +101,7 @@ Predicates 3D
 
 
 Transformations
-===============
-
-**Classes**
-
-.. autosummary::
-    :toctree: generated/
-    :nosignatures:
-
-    Projection
-    Reflection
-    Rotation
-    Scale
-    Shear
-    Transformation
-    Translation
-
-**Functions**
+---------------
 
 .. autosummary::
     :toctree: generated/
@@ -226,7 +168,7 @@ Transformations
 
 
 Linear algebra
-==============
+--------------
 
 .. autosummary::
     :toctree: generated/
@@ -275,7 +217,7 @@ Linear algebra
 
 
 Points, Vectors, Lines, Planes
-==============================
+------------------------------
 
 .. autosummary::
     :toctree: generated/
@@ -311,7 +253,7 @@ Points, Vectors, Lines, Planes
 
 
 Polygons & Polyhedrons
-======================
+----------------------
 
 .. autosummary::
     :toctree: generated/
@@ -335,7 +277,7 @@ Polygons & Polyhedrons
 
 
 Point Sets
-==========
+----------
 
 .. autosummary::
     :toctree: generated/
@@ -352,7 +294,7 @@ Point Sets
 
 
 Distance
-========
+--------
 
 .. autosummary::
     :toctree: generated/
@@ -370,7 +312,7 @@ Distance
 
 
 Intersections
-=============
+-------------
 
 .. autosummary::
     :toctree: generated/
@@ -399,7 +341,7 @@ Intersections
 
 
 Offsets
-=======
+-------
 
 .. autosummary::
     :toctree: generated/
@@ -411,7 +353,7 @@ Offsets
 
 
 Interpolation
-=============
+-------------
 
 .. autosummary::
     :toctree: generated/
@@ -424,7 +366,7 @@ Interpolation
 
 
 Fitting
-=======
+-------
 
 .. autosummary::
     :toctree: generated/
@@ -436,7 +378,7 @@ Fitting
 
 
 Boolean operations
-==================
+------------------
 
 .. autosummary::
     :toctree: generated/
@@ -448,7 +390,7 @@ Boolean operations
 
 
 Triangulation
-=============
+-------------
 
 .. autosummary::
     :toctree: generated/
@@ -463,7 +405,7 @@ Triangulation
 
 
 Triangle meshes
-===============
+---------------
 
 .. autosummary::
     :toctree: generated/
@@ -482,7 +424,7 @@ Triangle meshes
 
 
 Quad meshes
-===========
+-----------
 
 .. autosummary::
     :toctree: generated/
@@ -492,24 +434,13 @@ Quad meshes
 
 
 Pointclouds
-===========
+-----------
 
 .. autosummary::
     :toctree: generated/
     :nosignatures:
 
     icp_numpy
-
-
-Base Classes
-============
-
-.. autosummary::
-    :toctree: generated/
-    :nosignatures:
-
-    Primitive
-    Shape
 
 """
 from __future__ import absolute_import