+++ src/compas_rhino/objects/__init__.py
@@ -49,54 +49,6 @@ VolMeshObject
     :members: clear, draw, select_vertices, select_faces, select_edges, modify_vertices, modify_faces, modify_edges
     :no-show-inheritance:
 
-----
-
-Selecting
-=========
-
-.. autosummary::
-    :toctree: generated/
-    :nosignatures:
-
-    network_select_node
-    network_select_nodes
-    network_select_edge
-    network_select_edges
-    mesh_select_vertex
-    mesh_select_vertices
-    mesh_select_face
-    mesh_select_faces
-    mesh_select_edge
-    mesh_select_edges
-
-
-Modifying
-=========
-
-.. autosummary::
-    :toctree: generated/
-    :nosignatures:
-
-    network_update_attributes
-    network_update_node_attributes
-    network_update_edge_attributes
-    network_move_node
-    mesh_update_attributes
-    mesh_update_vertex_attributes
-    mesh_update_face_attributes
-    mesh_update_edge_attributes
-    mesh_move_vertex
-    mesh_move_vertices
-    mesh_move_face
-
-
-Inspecting
-==========
-
-.. autosummary::
-    :toctree: generated/
-    :nosignatures:
-
 
 """
 from __future__ import absolute_import