+++ src/compas_rhino/objects/inspect/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import absolute_import
 
-from .inspectors import *  # noqa: F401 F403
+from .meshinspector import *  # noqa: F401 F403
 
 __all__ = [name for name in dir() if not name.startswith('_')]