+++ src/compas_rhino/objects/modify/__init__.py
@@ -1,8 +1,6 @@
 from __future__ import absolute_import
 
 
-# from .primitives import *  # noqa : F401 F403
-# from .shapes import *  # noqa : F401 F403
 from .datastructures import *  # noqa : F401 F403